refactor : 안 쓰는 코드 주석처리 및 삭제
안 쓰는 코드를 주석 처리 하거나 삭제함

#52, #55
